+++ CHANGELOG.md
@@ -4,7 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
-### 🔄 Changed
+### 🐞 Fixed
+- Video tracks for anonymous users not displaying
 
 # [1.0.2](https://github.com/GetStream/stream-video-swift/releases/tag/1.0.2)
 _May 17, 2024_